UPD: sanity check for valid glut windows
git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/Vista/trunk@29742 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VistaCoreLibs/VistaDeviceDrivers/VistaDTrackDriver/VistaDTrackDriver.cpp
@@ -518,7 +518,7 @@ public:
 		pDeSer->ReadDelimitedString(strBeg, '[');
 		pDeSer->ReadDelimitedString(strBlock, ']');
 
-		int nNumDoubles = vecOut.size() / sizeof(double);
+		int nNumDoubles = (int)( vecOut.size() / sizeof(double) );
 		double* pTarget = (double*)&vecOut[0];
 		if( VistaConversion::ArrayFromString<double>( strBlock, pTarget, nNumDoubles, ' ' ) == false )
 		{
@@ -666,7 +666,7 @@ public:
 		pDeSer->ReadDelimitedString(strBeg, '[');
 		pDeSer->ReadDelimitedString(strBlock, ']');
 
-		int nNumDoubles = vecOut.size() / sizeof(double);
+		int nNumDoubles = (int)( vecOut.size() / sizeof(double) );
 		double* pTarget = (double*)&vecOut[0];
 		if( VistaConversion::ArrayFromString<double>( strBlock, pTarget, nNumDoubles, ' ' ) == false )
 		{
